abstract store from todo store, move todo logic into service
